requirements.txt
annotated-types==0.7.0
anyio==4.6.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
filelock==3.13.4
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.51.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
openai==1.48.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.2
pyswisseph==2.10.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.25.1
waitress==3.0.0
Werkzeug==3.0.2
